nueva

+++ crudRetos/views/vista_listar.php
@@ -9,22 +9,34 @@
             <th>OPCIONES</th>
         </tr>
         <?php
-            foreach ($dataToView["data"] as $datos) {
-                echo "<tr>
-                        <td>".$datos['id']."</td>
-                        <td>
-                            ".$datos['nombre']."
-                        </td>
-                        <td>
-                            <a href=\"index.php?action=editarCategoria&id=".$datos['id']."\">
-                             <span class=\"material-symbols-outlined\">edit_note</span>
-                            </a>
-                            <a href=\"index.php?action=eliminarCategoria&id=".$datos['id']."\">
-                             <span class=\"material-symbols-outlined\">delete</span>
-                            </a>
-                        </td>
-                    </tr>";
+        var_dump( $dataToView["data"]);
+            if($dataToView["data"]==''){  
+                foreach ($dataToView["data"] as $datos) {
+                    echo "<tr>
+                            <td>".$datos['id']."</td>
+                            <td>
+                                ".$datos['nombre']."
+                            </td>
+                            <td>
+                                <a href=\"index.php?action=editarCategoria&id=".$datos['id']."\">
+                                 <span class=\"material-symbols-outlined\">edit_note</span>
+                                </a>
+                                <a href=\"index.php?action=eliminarCategoria&id=".$datos['id']."\">
+                                 <span class=\"material-symbols-outlined\">delete</span>
+                                </a>
+                            </td>
+                        </tr>";
+                }
+                echo " </table>";
             }
+            else{
+                echo "</table>
+                <h1>No hay ninguna categoria creada pulse añadir para empezar a crearlas</h1>
+
+                ";
+                
+            }
+
         ?>
-    </table>
+   
 </main>
